[supernova] store packed in a react context and sliced in files
