+++ apps/supernova/src/components/alerts/shared/AlertSilencesList.jsx
@@ -1,19 +1,25 @@
 import React from "react"
 import { DateTime } from "luxon"
 
-import { Badge, DataGrid, DataGridCell, DataGridHeadCell, DataGridRow } from "juno-ui-components"
+import {
+  Badge,
+  DataGrid,
+  DataGridCell,
+  DataGridHeadCell,
+  DataGridRow,
+} from "juno-ui-components"
 
-import { useSilencesActions } from "../../../hooks/useStore"
+import { useSilencesActions } from "../../../hooks/useAppStore"
 
-const AlertSilencesList = ({alert}) => {
+const AlertSilencesList = ({ alert }) => {
   const dateFormat = { ...DateTime.DATETIME_SHORT }
   const timeFormat = { ...DateTime.TIME_24_WITH_SHORT_OFFSET }
-  
+
   const formatDateTime = (timestamp) => {
     const time = DateTime.fromISO(timestamp)
     return time.toLocaleString(dateFormat)
   }
-  
+
   const { getMappingSilences, getExpiredSilences } = useSilencesActions()
 
   const activeSilences = getMappingSilences(alert)
@@ -22,7 +28,7 @@ const AlertSilencesList = ({alert}) => {
 
   return (
     <>
-      { (silenceList.length > 0) &&
+      {silenceList.length > 0 && (
         <>
           <h2 className="text-xl font-bold mb-2 mt-8">Silences</h2>
           <DataGrid columns={5}>
@@ -33,25 +39,28 @@ const AlertSilencesList = ({alert}) => {
               <DataGridHeadCell>Created by</DataGridHeadCell>
               <DataGridHeadCell>Comment</DataGridHeadCell>
             </DataGridRow>
-          { silenceList.map((silence) => (
-            <DataGridRow key={silence.id}>
-              <DataGridCell>
-                <div>
-                  <Badge 
-                    variant={silence.status?.state === "active" ? "info" : "default"}>
+            {silenceList.map((silence) => (
+              <DataGridRow key={silence.id}>
+                <DataGridCell>
+                  <div>
+                    <Badge
+                      variant={
+                        silence.status?.state === "active" ? "info" : "default"
+                      }
+                    >
                       {silence.status?.state}
                     </Badge>
-                </div>
-              </DataGridCell>
-              <DataGridCell>{formatDateTime(silence.startsAt)}</DataGridCell>
-              <DataGridCell>{formatDateTime(silence.endsAt)}</DataGridCell>
-              <DataGridCell>{silence.createdBy}</DataGridCell>
-              <DataGridCell>{silence.comment}</DataGridCell>
-            </DataGridRow>
-          ))}
+                  </div>
+                </DataGridCell>
+                <DataGridCell>{formatDateTime(silence.startsAt)}</DataGridCell>
+                <DataGridCell>{formatDateTime(silence.endsAt)}</DataGridCell>
+                <DataGridCell>{silence.createdBy}</DataGridCell>
+                <DataGridCell>{silence.comment}</DataGridCell>
+              </DataGridRow>
+            ))}
           </DataGrid>
         </>
-      }
+      )}
     </>
   )
 }
