+++ apps/supernova/src/components/filters/PredefinedFilters.jsx
@@ -1,8 +1,11 @@
-import React, {useState} from "react"
+import React, { useState } from "react"
 
 import { Tabs, TabList, Tab, TabPanel } from "juno-ui-components"
-import { useActivePredefinedFilter, useFilterActions, usePredefinedFilters } from "../../hooks/useStore"
-
+import {
+  useActivePredefinedFilter,
+  useFilterActions,
+  usePredefinedFilters,
+} from "../../hooks/useAppStore"
 
 const PredefinedFilters = () => {
   const { setActivePredefinedFilter } = useFilterActions()
@@ -10,8 +13,12 @@ const PredefinedFilters = () => {
   const activePredefinedFilter = useActivePredefinedFilter()
 
   // find the index of the active predefined filter in the list of predefined filters and set it as the selected index
-  const activePredefinedFilterIndex = predefinedFilters.findIndex((filter) => filter.name === activePredefinedFilter)
-  const [selectedIndex, setSelectedIndex] = useState(activePredefinedFilterIndex)
+  const activePredefinedFilterIndex = predefinedFilters.findIndex(
+    (filter) => filter.name === activePredefinedFilter
+  )
+  const [selectedIndex, setSelectedIndex] = useState(
+    activePredefinedFilterIndex
+  )
 
   const handleTabSelect = (index) => {
     setSelectedIndex(index)
@@ -19,15 +26,19 @@ const PredefinedFilters = () => {
   }
 
   return (
-    <Tabs selectedIndex={selectedIndex} onSelect={(index) => handleTabSelect(index)} className="mb-4">
+    <Tabs
+      selectedIndex={selectedIndex}
+      onSelect={(index) => handleTabSelect(index)}
+      className="mb-4"
+    >
       <TabList>
-      { predefinedFilters.map((filter) => 
-        <Tab key={filter.name}>{filter.displayName}</Tab>
-      )}
+        {predefinedFilters.map((filter) => (
+          <Tab key={filter.name}>{filter.displayName}</Tab>
+        ))}
       </TabList>
-      { predefinedFilters.map((filter) => 
+      {predefinedFilters.map((filter) => (
         <TabPanel key={filter.name}></TabPanel>
-      )}
+      ))}
     </Tabs>
   )
 }
