+++ apps/supernova/src/components/silences/SilenceNew.jsx
@@ -17,7 +17,7 @@ import {
   useGlobalsApiEndpoint,
   useSilencesActions,
   useAlertEnrichedLabels,
-} from "../../hooks/useStore"
+} from "../../hooks/useAppStore"
 import { post } from "../../api/client"
 import AlertDescription from "../alerts/shared/AlertDescription"
 import SilenceNewAdvanced from "./SilenceNewAdvanced"
