+++ apps/supernova/src/lib/createAuthDataSlice.js
@@ -1,0 +1,57 @@
+export const AUTH_ACTIONS = {
+  SIGN_ON: "signOn",
+  SIGN_OUT: "signOut",
+}
+
+const createAuthDataSlice = (set, get) => ({
+  auth: {
+    data: null,
+    isProcessing: false,
+    loggedIn: false,
+    error: null,
+    lastAction: {},
+    appLoaded: false,
+    appIsLoading: false,
+
+    actions: {
+      setAppLoaded: (appLoaded) => {
+        set(
+          (state) => ({ auth: { ...state.auth, appLoaded } }),
+          false,
+          "auth/setAppLoaded"
+        )
+      },
+      setData: (data) => {
+        if (!data) return
+        set(
+          (state) => ({
+            auth: {
+              ...state.auth,
+              isProcessing: data?.isProcessing,
+              loggedIn: data?.loggedIn,
+              error: data?.error,
+              data: data?.auth,
+            },
+          }),
+          false,
+          "auth/setData"
+        )
+      },
+      setAction: (name) =>
+        set(
+          (state) => ({
+            auth: {
+              ...state.auth,
+              lastAction: { name: name, updatedAt: Date.now() },
+            },
+          }),
+          false,
+          "auth/setAction"
+        ),
+      login: () => get().auth.actions.setAction(AUTH_ACTIONS.SIGN_ON),
+      logout: () => get().auth.actions.setAction(AUTH_ACTIONS.SIGN_OUT),
+    },
+  },
+})
+
+export default createAuthDataSlice
