+++ apps/supernova/src/lib/createUserActivitySlice.js
@@ -1,0 +1,21 @@
+import produce from "immer"
+
+const createUserActivitySlice = (set, get) => ({
+  userActivity: {
+    isActive: true,
+
+    actions: {
+      setIsActive: (activity) => {
+        set(
+          (state) => ({
+            userActivity: { ...state.userActivity, isActive: activity },
+          }),
+          false,
+          "userActivity.setIsActive"
+        )
+      },
+    },
+  },
+})
+
+export default createUserActivitySlice
