+++ apps/supernova/src/lib/fakeObjects.js
